get2gnow #0000FCaEC - Whatfore & wherehow...Worth while stuff & a bunch of wtf
I'd finally & completely fixed my many times aforementioned issues with get2gnow appending twitter's latest update id over & over.  Its a gconf sync issue, gconf client itsn't saving the newest id correctly for timelines & replies.  So it was appending the latest update over & over & over & over & over & over & over & again...

& yes I want to keep typing over & over - over & over.  Its almost as annoying as the bug felt, but not as annoying as the fix.  I do know its fixed: finally, fully, & completely.  What I still don't get is why it wasn't breaking in more placese.  ~shrug~ whatfore & wherehow.  ;-) yeah, I just wanted to say it, in fact I feel I may have a release name.

*uber-w00t what I've finally figured it out what & why get2gnow was endlessly appending the last new twitter update endlessly & so *uber-w00t!* I've finally fixed get2gnow's since_id issue.  It was a problem with casting gulong to gfloat.  Once I realized, thnx to many many egreps through get2gnow's debug log,  I've had to refactor my User & UserStatus objects to parse user ids as floats and refactored my OnlineService methods for requesting last update ids to use float values.  I'd already added get & set float methods to my gconf client aspect.  I'm using float values becuase I have very lil doubt that status ids will soon out grow the maximum integer size, especially on smaller machines like netbooks and etc.

I really need to think about using sqlite or something else for storage so I can have something more flexible & platform indipenant - but that's a much longer term goal...well its at least something for my next version.

Now on to stuff some new stuff.

I've refactored, yes again, my debugging aspect.  Its now much more efficiant and has a much lower memory foot print.  I continue to improve my debugging aspect because of how much get2gnow uses this system & how much data is either out put or logged.  So I really want this system to be as advanced as possible.

Thnx to identi.ca's outage at I've also extended support for my networking callback methods.  Now my replacement for libsoup's callback is expected to return a 'void *' & recieves an OnlineServiceWrapper object as its 3rd parameter.  My after soup callback now recieves 2 parameters, 1st is the same OnlineServiceWrapper mentioned above.  It will only 'run' its callback methods once but otherwise all data from the request is accessible to the callback method.

I've also refactored how tweet-list sets its timeout to improve loading speeds.  Now it only calculates timeout seconds when initially loading & setting up the TweetList.  It will recalculate how long to wait between refreshing timelines

I've also improved its appearance an tooltips by implement SexyTreeView instead of GtkTreeView. I'm still using GtkBuilder files & still using Glade-3 to design my UIs but I've added ex commands as part of make that simply converts 'data/tweet-list.in.ui' to 'data/tweet-list.ui' while converting objects of GtkTreeView class to SexyTreeView.  I plan on implementin this in all of my GtkBuilder ui files, but for now its just implemented on in TweetList.

I've also improve my date method for parsing the dates of when an update was posted & its age, eg 1, 2, 3, etc seconds, minutes, hours, ago.  As part of that I've fixed all previous issues with how get2gnow was update the 'age' of old updates when new updates are appended to the TweetList.  I was having it use GTK's GtkTreeModelForEachFunc and it just wasn't working so instead I just wrote my own method that updates the dates of all entries in the list store.

So yeah its another big update with tons of stuff.  Stuff for users...stuff for artist...stuff for hackers... & especially stuff those uber chick nerds who are all three.  Aren't we special :-)...why yes we are, most certainly.

Committed on: 2009-06-26 @ 12:27:19 AM

+++ src/online-service-request.c
@@ -117,8 +117,8 @@ typedef struct _OnlineServiceRequestPopup OnlineServiceRequestPopup;
 
 struct _SelectedTweet{
 	OnlineService	*service;
-	gulong		id;
-	gulong		user_id;
+	gfloat		id;
+	gfloat		user_id;
 	gchar		*user_name;
 	gchar		*tweet;
 };
@@ -423,11 +423,11 @@ static void online_service_request_free(OnlineServiceRequest *request){
 /********************************************************************************
  *               selected_tweet methods, handlers, callbacks, & etc.            *
  ********************************************************************************/
-void set_selected_tweet(OnlineService *service, const gulong id, const gulong user_id, const gchar *user_name, const gchar *tweet){
-	/*	gulong id=strtoul(string, NULL, 10);	*/
+void set_selected_tweet(OnlineService *service, const gfloat id, const gfloat user_id, const gchar *user_name, const gchar *tweet){
+	/*	gint id=atoi(string);	*/
 	if(selected_tweet) unset_selected_tweet();
 	
-	debug("SelectedTweet created from '%s', tweet id: #%lu from '%s' on '%s'.", online_service_get_key(service), id, user_name, online_service_get_uri(service));
+	debug("SelectedTweet created from '%s', tweet id: #%f from '%s' on '%s'.", online_service_get_key(service), id, user_name, online_service_get_uri(service));
 	selected_tweet=g_new0(SelectedTweet, 1);
 	selected_tweet->service=service;
 	selected_tweet->id=id;
@@ -440,7 +440,7 @@ OnlineService *selected_tweet_get_service(void){
 	return ( (selected_tweet && selected_tweet->service) ?selected_tweet->service :NULL );
 }/*selected_tweet_get_service();*/
 
-gulong selected_tweet_get_id(void){
+gfloat selected_tweet_get_id(void){
 	return ( (selected_tweet && selected_tweet->id) ?selected_tweet->id :0 );
 }/*selected_tweet_get_user_id();*/
 
@@ -448,7 +448,7 @@ gchar *selected_tweet_get_user_name(void){
 	return ( (selected_tweet && selected_tweet->user_name) ?selected_tweet->user_name :NULL );
 }/*selected_tweet_get_user_name();*/
 
-gulong selected_tweet_get_user_id(void){
+gfloat selected_tweet_get_user_id(void){
 	return ( (selected_tweet && selected_tweet->user_id) ?selected_tweet->user_id :0 );
 }/*selected_tweet_get_user_id();*/
 
@@ -530,14 +530,14 @@ void online_service_request_selected_tweet_unblock(void){
 
 void online_service_request_selected_tweet_save_fave(void){
 	if(!(selected_tweet && selected_tweet->id)) return;
-	gchar *fave_tweet_id=g_strdup_printf( "%lu", selected_tweet->id );
+	gchar *fave_tweet_id=g_strdup_printf( "%f", selected_tweet->id );
 	online_service_request_main(selected_tweet->service, Fave, ( gconfig_if_bool(PREFS_TWEET_VIEW_DIALOG, FALSE) ?tweet_view_get_window() :main_window_get_window() ), fave_tweet_id);
 	g_free(fave_tweet_id);
 }/*online_service_request_selected_tweet_save_fave*/
 
 void online_service_request_selected_tweet_destroy_fave(void){
 	if(!(selected_tweet && selected_tweet->id)) return;
-	gchar *fave_tweet_id=g_strdup_printf( "%lu", selected_tweet->id );
+	gchar *fave_tweet_id=g_strdup_printf( "%f", selected_tweet->id );
 	online_service_request_main(selected_tweet->service, UnFave, ( gconfig_if_bool(PREFS_TWEET_VIEW_DIALOG, FALSE) ?tweet_view_get_window() :main_window_get_window() ), fave_tweet_id);
 	g_free(fave_tweet_id);
 }/*online_service_request_selected_tweet_destroy_fave*/
